feat: complete transaction form refactor
