+++ src/components/index.js
@@ -2,5 +2,4 @@ export { default as UniversalTable } from "./UniversalTable/UniversalTable";
 export { default as ModalWindow } from "./ModalWindow/ModalWindow";
 export { default as UniversalButton } from "./UniversalButton/UniversalButton";
 export { default as WalletsForm } from "./WalletsForm/WalletsForm";
-// export { default as TransactionForm } from "./TransactionForm/TransactionForm";
-export { default as TransactionForm } from "./TransactionForm/TransactionForm__experimental";
+export { default as TransactionForm } from "./TransactionForm/TransactionForm";
